+++ src/Pwa/setHideConfirm.mjs
@@ -1,5 +1,3 @@
-/** @typedef {import("./hideConfirm.mjs").hideConfirm} hideConfirm */
-
 /**
- * @typedef {(hide_confirm: hideConfirm) => void} setHideConfirm
+ * @typedef {(hide_confirm: () => void) => void} setHideConfirm
  */